server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Bcrypt==1.0.1
marshmallow==3.19.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0
alembic==1.15.2